requirements.txt
grpcio ~= 1.50.0